chore(api): extract pg-error helper + log err.cause in lineups matching (ROK-1225)
Move `extractErrorDetail` from `itad-price-sync.service.ts` to a shared
`api/src/common/pg-error.helpers.ts` so any catcher of a Drizzle-wrapped
postgres error can surface code/detail/hint instead of the opaque
"Failed query: …" string. Re-export from the original location to keep
the existing adversarial test suite passing without churn.

Wire the helper into `runMatchingAlgorithm` (lineups-lifecycle.helpers.ts:144)
— the catch block was logging only `err.message`, which is the wrapped
"Failed query" prefix. Future masked failures (FK violations,
unique-constraint collisions) now hit the log with their real PG code so we
don't have to re-instrument when the next regression lands.

Load-bearing for ROK-1225 root-cause diagnosis.
